Separate LyapunovContinuousTimeHybridSystem to its own file. Currently the file is too big.
